# Use a giant UBO to optimize performance in 2D [OpenGL3]