… drivers/gles3/rasterizer_canvas_gles3.cpp